requirements.txt
asgiref==3.4.0
DateTime==4.3
dj-database-url==0.5.0
Django==3.2.4
django-appconf==1.0.4
django-heroku==0.0.0
django-staticfiles==1.2.1
djangorestframework==3.12.4
gunicorn==20.1.0
psycopg2-binary==2.9.1
pytz==2021.1
sqlparse==0.4.1
whitenoise==5.2.0
zope.interface==5.4.0
